Refactor job comparision business case
